backend/pyproject.toml
[project]
name = "archon-backend"
version = "0.1.0"
description = "FastAPI + LangGraph backend for Archon"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.0",
  "email-validator>=2.3.0",
  "sqlalchemy[asyncio]>=2.0.43",
  "asyncpg>=0.30.0",
  "alembic>=1.16.0",
  "httpx>=0.28.1",
  "beautifulsoup4>=4.13.5",
  "lxml>=6.0.1",
  "openai>=1.102.0",
  "langchain>=0.3.27",
  "langchain-openai>=0.3.33",
  "langgraph>=0.6.6",
  "google-api-python-client>=2.181.0",
  "google-auth>=2.40.3",
  "google-auth-oauthlib>=1.2.2",
  "python-dotenv>=1.1.1",
  "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.2",
  "pytest-asyncio>=1.1.0",
  "ruff>=0.12.11",
  "mypy>=1.17.1",
  "types-aiofiles>=24.1.0.20250822",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "ASYNC", "B"]
ignore = ["E501", "B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
